Update patch set 2
Patch Set 2:

I need to check if this is a real limitation and expand the job to accept other branches. 
For now, I do not have permissions to overwrite a +1 due to the Prolog rules in place. But let me force a verify on this change.. let me see if I can do that.

Patch-set: 2
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>