Update patch set 2
Patch Set 2:

Ok.. so I was checking further this request. The committer permissions for a particular repo cannot be adjusted depending of a branch of a repo. What we have internally in the access permissions of a repo is just an "Owner" permission for "refs/*". We do not support specific permissions for "refs/heads/<branch>" for branches in specific. Neither we have support for separate LDAP groups depending on the branch. In case of ONAP, this is not maintainable as we would have to have "n" number of groups (for each branch) for each repo (200 repos aprox) which is not easy to maintain internally. 
 

You will need to make this change against the master branch please.

Patch-set: 2
Reviewer: Gerrit User 481 <481@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>