Update patch set 1
Patch Set 1: -Code-Review

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > @Shashikanth VH what is your opinion on Morgan's comment?
> 
> vfc-huawei-vnmf-driver is required, we further look into other possibilities in next releases.

ok, I removed my -1
do you have any automated tests / usecases using this pod?
the fact to deploy it as an addons does not mean that we removed it. So it could be upload on demand if needed and not necessarily on the main installation.

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0